more precision to highlow

+++ RCSwitch.h
@@ -109,8 +109,8 @@ class RCSwitch {
      * Thus, the pulse overall lasts (high+low)*pulseLength
      */
     struct HighLow {
-        uint8_t high;
-        uint8_t low;
+        uint16_t high;
+        uint16_t low;
     };
 
     /**
